+++ R/dt.R
@@ -3,7 +3,7 @@
 #' @description
 #' Pipeable data.table call
 #'
-#' @param x See ?data.table:::`[.data.table`
+#' @param .data A data.frame or data.table
 #' @param i See ?data.table:::`[.data.table`
 #' @param j See ?data.table:::`[.data.table`
 #' @param by See ?data.table:::`[.data.table`
@@ -31,4 +31,27 @@
 #'   as_dt() %>%
 #'   dt(, ':='(double_x = x * 2)) %>%
 #'   dt(order(-double_x))
-dt <- data.table:::`[.data.table`
+dt <- function(.data, i, j, by, keyby, with = TRUE,
+               nomatch = getOption("datatable.nomatch", NA),
+               mult = "all",
+               roll = FALSE,
+               rollends = if (roll=="nearest") c(TRUE,TRUE)
+               else if (roll>=0) c(FALSE,TRUE)
+               else c(TRUE,FALSE),
+               which = FALSE,
+               .SDcols,
+               verbose = getOption("datatable.verbose"),                   # default: FALSE
+               allow.cartesian = getOption("datatable.allow.cartesian"),   # default: FALSE
+               drop = NULL, on = NULL) {
+
+  if (!is.data.frame(.data)) stop(".data must be a data.frame or data.table")
+
+  sys_call <- sys.call()
+
+  sys_call[[1]] <- as.symbol("[")
+
+  if (!is.data.table(.data))
+    sys_call[[2]] <- substitute(as.data.table(.data))
+
+  eval_tidy(sys_call, env = caller_env())
+}