+++ R/select_helpers.R
@@ -24,7 +24,7 @@
 #'   dt_select(dt_starts_with("x"))
 #'
 #' example_dt %>%
-#' dt_select(dt_ends_with("y"))
+#'   dt_select(dt_ends_with("y"))
 dt_starts_with <- function(match) {
   .names <- names(caller_env())
   .names <- .names[!.names %in% c("is.integer", "is.double", "is.numeric",