+++ R/tidytable_package.R
@@ -1,6 +1,6 @@
 # Suppress R CMD check note
 #' @import data.table
-#' @importFrom data.table data.table melt dcast
+#' @importFrom data.table as.data.table copy data.table is.data.table melt dcast
 #' @importFrom rlang abort caller_env enexpr enexprs expr eval_tidy
 #' @importFrom rlang is_formula is_named parse_expr seq2 sym syms
 #' @importFrom stats as.formula na.omit setNames