+++ NEWS.md
@@ -1,6 +1,9 @@
 # tidytable 0.3.1 (in development)
 
+* CRAN maintenance update with a couple small changes added on
 * `dt_left_join()` fix so columns are ordered correctly
+* Add group by ability to `dt_mutate_across()`
+* `dt()` now automatically converts data.frame input to a data.table
 
 # tidytable 0.3.0